requirements.txt
fastapi
uvicorn
langchain-mistralai
langchain-core
python-dotenv